Move Websocket methods to Cortex and handlers to Headset
